pyproject.toml
[tool.poetry]
name = "groq-ai"
version = "0.1.0"
description = ""
authors = ["Ricardo Roson <rroson@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
groq = "^0.9.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
